added nodeinfo logic to oe shim
